+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nrfcloud/cdktf-aws-adaptor
 
+## 0.7.1
+
+### Patch Changes
+
+- 8a6b195: fix issue with inverted dependency synthesis
+
+## 0.7.0
+
+### Minor Changes
+
+- 5bd0311: Add support for Iam AccessKey resource mapping
+- a2381fa: Fix for iam policy eventual consistency
+
 ## 0.6.2
 
 ### Patch Changes